fix: add the logout button
Closes: https://github.com/alemayhu/2anki.net/issues/532
