+++ web/src/components/NavigationBar/helpers/getCookie.ts
@@ -1,0 +1,15 @@
+// https://www.w3schools.com/js/js_cookies.asp
+export default function getCookie(cname) {
+  const name = `${cname}=`;
+  const ca = document.cookie.split(';');
+  for (let i = 0; i < ca.length; i += 1) {
+    let c = ca[i];
+    while (c.charAt(0) === ' ') {
+      c = c.substring(1);
+    }
+    if (c.indexOf(name) === 0) {
+      return c.substring(name.length, c.length);
+    }
+  }
+  return '';
+}
